Add lots more rebraining notes
Here I included photos of quite a number of items available on
MercadoLibre, with commentary on what their dissected organs might be
reusable for.  This involved some clarification to LICENSE.  In passing
I also added a little text to the taking-screenshots note, reformatted
the garden-light-panel note, and added some notes about piezo buzzers to
the minimal-cost-computer note.