Add janky inline image location in HTML dir
If I put the images in liabilities/ in the source, they weren’t going to
show up when I edited the Markdown, or in GitLab’s view.  So I wanted to
put them in markdown/.  Then I still was faced with the choice of
copying them to liabilities/ in the output, or into notes/.  If I copy
them into liabilities/ I need to fiddle with the inline image tags in
the generated HTML to fix up the references.  So instead I’m just
copying them over, one by one, when I run.